compilation warnings
git-svn-id: svn+ssh://kadu.net/home/kadu/repos/kadu/trunk@6329 fea5205d-6df5-0310-bef7-ea67e47da9a5

+++ kadu-core/kadu_parser.cpp
@@ -94,22 +94,27 @@ bool KaduParser::unregisterObjectTag(const QString &name, ObjectTagCallback)
 QString KaduParser::executeCmd(const QString &cmd)
 {
 	kdebugf();
+
 	QString s(cmd);
 	s.remove(QRegExp("`|>|<"));
-	s.append(" >");
-	s.append(ggPath("execoutput"));
+	s.append(" > " + ggPath("execoutput"));
 
-	system(qPrintable(s));
-	QFile *f = new QFile(ggPath("execoutput"));
-	if (f->open(QIODevice::ReadOnly))
+	int ret = system(qPrintable(s));
+
+	s = QString::null;
+
+	if (ret != -1)
 	{
-		s = QString(f->readAll());
-		f->close();
-		QFile::remove(ggPath("execoutput"));
+		QFile *f = new QFile(ggPath("execoutput"));
+		if (f->open(QIODevice::ReadOnly))
+		{
+			s = QString(f->readAll());
+			f->close();
+			QFile::remove(ggPath("execoutput"));
+		}
+		delete f;
 	}
-	else
-		s = QString::null;
-	delete f;
+
 	kdebugf2();
 	return s;
 }
